:memo: Rename itemroot -> item
