+++ @navikt/core/react/src/form/file-upload/item/FileUploadItem.types.ts
